Capsule impl removed and remaining moved to `circles.rs`
